config: split file.c for CONFIG_TYPE_FILE related functions
